requirements.txt
sons
regex
colorama
GitPython